[TE] frontend - harleyjj/full-ui - form and yaml share state except rules
