+++ thirdeye/thirdeye-frontend/app/styles/components/subscription-yaml.scss
@@ -1,5 +1,5 @@
-.yaml-editor {
-  .yaml-editor-msg {
+.subscription-yaml {
+  .subscription-yaml-msg {
     color: #e75858;
 
     &__icon {
